+++ lib/isolate.dart
@@ -22,10 +22,16 @@ class IsolateRef {
 
   const IsolateRef._internal(this._isolate, this._sendPort, this._properties);
 
+  /**
+   * Sends a message to the isolate represented by this reference.
+   */
   send(String message) {
     _sendPort.send(new _PayloadMsg(message));
   }
 
+  /**
+   * Returns the value of a property of the represented isolate.
+   */
   property(String key) {
     return _properties[key];
   }
@@ -33,7 +39,8 @@ class IsolateRef {
 }
 
 /**
- * This context is provided to the entry point of an isolate spawned by IsolateCluster.
+ * This context is provided to the [EntryPoint] of an isolate spawned by [IsolateCluster]. A context is bound to a single
+ * isolate.
  */
 class IsolateContext {
 
@@ -49,20 +56,39 @@ class IsolateContext {
     _receivePort.listen((msg) => _processMessage(msg));
   }
 
+
+  /**
+   * Returns the value of a property of the isolate this context is bound to.
+   */
   property(String key) => _properties[key];
 
+  /**
+   * A stream of message sent to the isolate this context is bound to.
+   */
   Stream<String> get onMessage => _payloadStreamController.stream;
 
+  /**
+   * A stream of isolate up events.
+   */
   Stream<IsolateRef> get onIsolateUp => _isolateUpStreamController.stream;
 
+  /**
+   * The [ShutdownRequestListener] is called, when the isolate this context is bound to, receives a shutdown request.
+   */
   set shutdownRequestListener(ShutdownRequestListener listener) => _shutdownRequestListener = listener;
 
+  /**
+   * Shuts down the isolate this context is bound to.
+   */
   shutdownIsolate() {
     _payloadStreamController.close();
     _isolateUpStreamController.close();
     _sendPort.send(_IsolateReadyForShutdownMsg.INSTANCE);
   }
 
+  /**
+   * Shuts down the node of the isolate this context is bound to.
+   */
   shutdownNode({Duration timeout}) {
     _sendPort.send(new _NodeShutdownRequestMsg(timeout));
   }